Require claims to be signed with subject address and claim type
This makes claims valid only for the intended recipient and claim type.

The prefix stuff here is being done to accomodate the prefixing done by
web3 when signing messages, as a security measure
(https://github.com/ethereum/go-ethereum/issues/3731). Because this
signing is most likely going to be done on servers of major claim
validators rather than from end-user clients, the web3 prefix could
probably be bypassed. But to keep things simple I just adjusted the
solidity code to use this prefix. I think it's adequate for now.